# api/openapi.yaml
openapi: 3.0.0
info:
  title: API de particulares sancionados
  version: 1.1.0
...
servers:
- url: /
paths:
  /v1/psancionados:
    post:
      tags:
      - psancionados
      summary: Muestra los particulares sancionados permitiendo búsquedas avanzadas.
      operationId: post_psancionados
      requestBody:
        description: JSON para peticiones de busqueda avanzada.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reqParticularesSancionados'
        required: true
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resParticularesSancionados'
        "400":
          description: Página inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
        "401":
          description: No tienes permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
      security:
      - psancionados_auth: []
      x-swagger-router-controller: Psancionados
  /v1/psancionados/dependencias:
    get:
      tags:
      - psancionados
      summary: Dependencias donde se cometió la sanción
      operationId: get_dependencias
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dependencias'
        "400":
          description: Página inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
        "401":
          description: No tienes permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
      security:
      - psancionados_auth: []
      x-swagger-router-controller: Psancionados
components:
  schemas:
    resParticularesSancionados:
      required:
      - pagination
      - results
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/resParticularesSancionados_pagination'
        results:
          type: array
          items:
            $ref: '#/components/schemas/resParticularesSancionados_results'
      example:
        pagination:
          hasNextPage: true
          pageSize: 30
          page: 20
          totalRows: 300
        results:
        - autoridadSancionadora: OIC
          responsableSancion:
            primerApellido: Pérez
            segundoApellido: López
            nombres: Carlos
          institucionDependencia:
            clave: XYZ987
            siglas: SHCP
            nombre: Secretaría de Hacienda y Crédito Público
          tipoFalta: Contratación indebida
          resolucion:
            sentido: Sansionatoria con multa
            fechaNotificacion: 2019-11-20T00:00:00.000+0000
            url: https://ejemplo.mx/archivo/documento.pdf
          expediente: ABC123
          multa:
            monto: 5000000
            moneda:
              clave: MXN
              valor: Peso Mexicano
          tipoSancion:
          - clave: I
            valor: Inhabilitado
          - clave: M
            valor: Multado
          acto: Contrataciones
          documentos:
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_INHABILITACION
            titulo: Constancia de inhabilitación...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_INHABILITACION
            titulo: Constancia de inhabilitación...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          fechaCaptura: 2017-07-21T17:32:28Z
          particularSancionado:
            nombreRazonSocial: PATITO TELECOM S.A. de C.V.
            objetoSocial: Objeto social de la empresa
            domicilioExtranjero:
              ciudadLocalidad: Springfield
              codigoPostal: "01234"
              estadoProvincia: North Takoma
              calle: Evergreen Terrace
              numeroExterior: "742"
              numeroInterior: ""
              pais:
                clave: MX
                valor: México
            tipoPersona: F
            directorGeneral:
              primerApellido: Pérez
              segundoApellido: López
              curp: BADD110313HCMLNS09
              nombres: Carlos
            domicilioMexico:
              codigoPostal: "55018"
              municipio:
                clave: "033"
                valor: Ecatepec de Morelos
              entidadFederativa:
                clave: "15"
                valor: México
              localidad:
                clave: "0001"
                valor: Ecatepec de Morelos
              numeroExterior: "24"
              numeroInterior: "48"
              vialidad:
                clave: CALLE
                valor: El Rosal
              pais:
                clave: MX
                valor: México
            apoderadoLegal:
              primerApellido: Pérez
              segundoApellido: López
              curp: BADD110313HCMLNS09
              nombres: Carlos
            telefono: "552123244"
            rfc: PAT21343213
          causaMotivoHechos: Causa o motivo de la sanción
          observaciones: Se multo e inhabilitó al proveedor PATITO TELECOM S.A de
            C.V. por ...
          id: "1"
          objetoContrato: Servicio de telefonía fija
          inhabilitacion:
            fechaInicial: 2019-01-01T00:00:00.000+0000
            plazo: 2 años
            fechaFinal: 2020-12-31T00:00:00.000+0000
        - autoridadSancionadora: OIC
          responsableSancion:
            primerApellido: Pérez
            segundoApellido: López
            nombres: Carlos
          institucionDependencia:
            clave: XYZ987
            siglas: SHCP
            nombre: Secretaría de Hacienda y Crédito Público
          tipoFalta: Contratación indebida
          resolucion:
            sentido: Sansionatoria con multa
            fechaNotificacion: 2019-11-20T00:00:00.000+0000
            url: https://ejemplo.mx/archivo/documento.pdf
          expediente: ABC123
          multa:
            monto: 5000000
            moneda:
              clave: MXN
              valor: Peso Mexicano
          tipoSancion:
          - clave: I
            valor: Inhabilitado
          - clave: M
            valor: Multado
          acto: Contrataciones
          documentos:
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_INHABILITACION
            titulo: Constancia de inhabilitación...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_INHABILITACION
            titulo: Constancia de inhabilitación...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          fechaCaptura: 2017-07-21T17:32:28Z
          particularSancionado:
            nombreRazonSocial: PATITO TELECOM S.A. de C.V.
            objetoSocial: Objeto social de la empresa
            domicilioExtranjero:
              ciudadLocalidad: Springfield
              codigoPostal: "01234"
              estadoProvincia: North Takoma
              calle: Evergreen Terrace
              numeroExterior: "742"
              numeroInterior: ""
              pais:
                clave: MX
                valor: México
            tipoPersona: F
            directorGeneral:
              primerApellido: Pérez
              segundoApellido: López
              curp: BADD110313HCMLNS09
              nombres: Carlos
            domicilioMexico:
              codigoPostal: "55018"
              municipio:
                clave: "033"
                valor: Ecatepec de Morelos
              entidadFederativa:
                clave: "15"
                valor: México
              localidad:
                clave: "0001"
                valor: Ecatepec de Morelos
              numeroExterior: "24"
              numeroInterior: "48"
              vialidad:
                clave: CALLE
                valor: El Rosal
              pais:
                clave: MX
                valor: México
            apoderadoLegal:
              primerApellido: Pérez
              segundoApellido: López
              curp: BADD110313HCMLNS09
              nombres: Carlos
            telefono: "552123244"
            rfc: PAT21343213
          causaMotivoHechos: Causa o motivo de la sanción
          observaciones: Se multo e inhabilitó al proveedor PATITO TELECOM S.A de
            C.V. por ...
          id: "1"
          objetoContrato: Servicio de telefonía fija
          inhabilitacion:
            fechaInicial: 2019-01-01T00:00:00.000+0000
            plazo: 2 años
            fechaFinal: 2020-12-31T00:00:00.000+0000
    reqParticularesSancionados:
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/reqParticularesSancionados_sort'
        page:
          type: integer
          description: Número de página a mostrar, en caso de no recibir el parámetro,
            su valor será el indicado en default.
          example: 1
          default: 1
        pageSize:
          type: integer
          description: Número de registros a mostrar por página, el máximo permitido
            es 200 y el minimo 1, , en caso de no recibir el parámetro, su valor será
            el indicado en default.
          example: 10
          default: 10
        query:
          $ref: '#/components/schemas/reqParticularesSancionados_query'
        rfcSolicitante:
          type: string
          description: RFC de la persona quien hace la consulta, si se omite este
            parametro, se enviaran solo los datos públicos
          example: GOAP780710RH7
    dependencias:
      type: array
      description: Instituciones o dependencias donde se cometió la sanción
      items:
        $ref: '#/components/schemas/resParticularesSancionados_institucionDependencia'
    resError:
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Código alfa numérico del error de la consulta.
          example: abc123
        message:
          type: string
          description: Texto descriptivo para el usuario final del error en la consulta
          example: Error al formar la consulta, revise el formato de los campos.
      description: Estos valores son definidos por el sistema del API.
    tipoSancion:
      uniqueItems: true
      type: array
      description: Sanción impuesta
      example:
      - clave: I
        valor: Inhabilitado
      - clave: M
        valor: Multado
      items:
        $ref: '#/components/schemas/tipoSancion_inner'
...
    resParticularesSancionados_pagination:
      required:
      - hasNextPage
      - page
      - pageSize
      - totalRows
      type: object
      properties:
        pageSize:
          minimum: 1
          type: number
          description: Total de páginas
          example: 30
        page:
          minimum: 1
          type: number
          description: Número de página actual
          example: 20
        totalRows:
          minimum: 0
          type: number
          description: Total de registros que cumplen los parámetros de la petición
          example: 300
        hasNextPage:
          type: boolean
          description: Indica si existe otra página
          example: true
      example:
        hasNextPage: true
        pageSize: 30
        page: 20
        totalRows: 300
    resParticularesSancionados_institucionDependencia:
      required:
      - nombre
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: Secretaría de Hacienda y Crédito Público
        siglas:
          type: string
          description: Siglas de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: SHCP
        clave:
          type: string
          description: Clave de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: XYZ987
      example:
        clave: XYZ987
        siglas: SHCP
        nombre: Secretaría de Hacienda y Crédito Público
...
    resParticularesSancionados_particularSancionado_domicilioMexico:
      type: object
      properties:
        pais:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico_pais'
        entidadFederativa:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico_entidadFederativa'
        municipio:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico_municipio'
        codigoPostal:
          maxLength: 5
          type: string
          description: Código postal del domicilio
          example: "55018"
        localidad:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico_localidad'
        vialidad:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico_vialidad'
        numeroExterior:
          type: string
          description: Número exterior del domicilio
          example: "24"
        numeroInterior:
          type: string
          description: Número interior del domicilio
          example: "48"
      description: Dirección del contratista en caso de que se encuentre en México.
        Este dato es información confidencial
      example:
        codigoPostal: "55018"
        municipio:
          clave: "033"
          valor: Ecatepec de Morelos
        entidadFederativa:
          clave: "15"
          valor: México
        localidad:
          clave: "0001"
          valor: Ecatepec de Morelos
        numeroExterior: "24"
        numeroInterior: "48"
        vialidad:
          clave: CALLE
          valor: El Rosal
        pais:
          clave: MX
          valor: México
    resParticularesSancionados_particularSancionado_domicilioExtranjero:
      type: object
      properties:
        calle:
          type: string
          example: Evergreen Terrace
        numeroExterior:
          type: string
          example: "742"
        numeroInterior:
          type: string
          example: ""
        ciudadLocalidad:
          type: string
          example: Springfield
        estadoProvincia:
          type: string
          example: North Takoma
        pais:
          $ref: '#/components/schemas/pais'
        codigoPostal:
          type: string
          example: "01234"
      description: Llenar en caso de que el domicilio del particular se encuentre
        en el extranjero
      example:
        ciudadLocalidad: Springfield
        codigoPostal: "01234"
        estadoProvincia: North Takoma
        calle: Evergreen Terrace
        numeroExterior: "742"
        numeroInterior: ""
        pais:
          clave: MX
          valor: México
    resParticularesSancionados_particularSancionado_directorGeneral:
      type: object
      properties:
        nombres:
          type: string
          description: Nombre(s) del director general de la empresa al momento de
            la falta
          example: Carlos
        primerApellido:
          type: string
          description: Primer apellido del director general de la empresa al momento
            de la falta
          example: Pérez
        segundoApellido:
          type: string
          description: Segundo apellido del director general de la empresa al momento
            de la falta
          example: López
        curp:
          maxLength: 18
          type: string
          description: CURP del director general de la empresa al momento de la falta
          example: BADD110313HCMLNS09
      example:
        primerApellido: Pérez
        segundoApellido: López
        curp: BADD110313HCMLNS09
        nombres: Carlos
...
    resParticularesSancionados_particularSancionado:
      required:
      - nombreRazonSocial
      - tipoPersona
      type: object
      properties:
        nombreRazonSocial:
          type: string
          description: Razón social de la empresa
          example: PATITO TELECOM S.A. de C.V.
        objetoSocial:
          type: string
          description: Objeto social de la empresa
          example: Objeto social de la empresa
        rfc:
          type: string
          description: RFC del contratista.
          example: PAT21343213
        tipoPersona:
          type: string
          description: Tipo de persona. Valores permitidos F = Física, M = Moral
          example: F
          enum:
          - F
          - M
        telefono:
          maxLength: 12
          type: string
          description: Teléfono del contratista. Número teléfonico estandarizado.
            http://www.itu.int/dms_pub/itu-t/opb/sp/T-SP-E.164D-2009-PDF-S.pdf.
          example: "552123244"
        domicilioMexico:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioMexico'
        domicilioExtranjero:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_domicilioExtranjero'
        directorGeneral:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_directorGeneral'
        apoderadoLegal:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado_apoderadoLegal'
      example:
        nombreRazonSocial: PATITO TELECOM S.A. de C.V.
        objetoSocial: Objeto social de la empresa
        domicilioExtranjero:
          ciudadLocalidad: Springfield
          codigoPostal: "01234"
          estadoProvincia: North Takoma
          calle: Evergreen Terrace
          numeroExterior: "742"
          numeroInterior: ""
          pais:
            clave: MX
            valor: México
        tipoPersona: F
        directorGeneral:
          primerApellido: Pérez
          segundoApellido: López
          curp: BADD110313HCMLNS09
          nombres: Carlos
        domicilioMexico:
          codigoPostal: "55018"
          municipio:
            clave: "033"
            valor: Ecatepec de Morelos
          entidadFederativa:
            clave: "15"
            valor: México
          localidad:
            clave: "0001"
            valor: Ecatepec de Morelos
          numeroExterior: "24"
          numeroInterior: "48"
          vialidad:
            clave: CALLE
            valor: El Rosal
          pais:
            clave: MX
            valor: México
        apoderadoLegal:
          primerApellido: Pérez
          segundoApellido: López
          curp: BADD110313HCMLNS09
          nombres: Carlos
        telefono: "552123244"
        rfc: PAT21343213
    resParticularesSancionados_responsableSancion:
      required:
      - nombres
      - primerApellido
      type: object
      properties:
        nombres:
          type: string
          description: Nombre o nombres del Titular del área de responsabilidades
            o Contralor del Órgano Interno de Control responsable de la información
            registrada
          example: Carlos
        primerApellido:
          type: string
          description: Primer apellido del Titular del área de responsabilidades o
            Contralor del Órgano Interno de Control responsable de la información
            registrada
          example: Pérez
        segundoApellido:
          type: string
          description: Segundo apellido del Titular del área de responsabilidades
            o Contralor del Órgano Interno de Control responsable de la información
            registrada
          example: López
      example:
        primerApellido: Pérez
        segundoApellido: López
        nombres: Carlos
    resParticularesSancionados_resolucion:
      type: object
      properties:
        sentido:
          type: string
          description: Sentido de la resolución. Por ejemplo, Sancionatoria con multa.
          example: Sansionatoria con multa
        url:
          type: string
          description: URL que apunta al documento en formato digital de la resolución
            emitida por el correspondiente Organo Interno de Control
          example: https://ejemplo.mx/archivo/documento.pdf
        fechaNotificacion:
          type: string
          description: Fecha de notificación de la resolución. Fecha en formato ISO
            8601
          format: date
          example: 2019-11-20
      example:
        sentido: Sansionatoria con multa
        fechaNotificacion: 2019-11-20T00:00:00.000+0000
        url: https://ejemplo.mx/archivo/documento.pdf
...
    resParticularesSancionados_multa:
      type: object
      properties:
        monto:
          minimum: 0
          type: number
          description: Monto de la multa impuesta a la empresa
          format: float
          example: 5000000
        moneda:
          $ref: '#/components/schemas/resParticularesSancionados_multa_moneda'
      description: Se llena en caso de que el particular haya tenido multa económica
      example:
        monto: 5000000
        moneda:
          clave: MXN
          valor: Peso Mexicano
    resParticularesSancionados_inhabilitacion:
      type: object
      properties:
        plazo:
          type: string
          description: Plazo de la inhabilitación. Ej. 2 años
          example: 2 años
        fechaInicial:
          type: string
          description: Fecha inicial de la inhabilitación. Fecha en formato ISO 8601
          format: date
          example: 2019-01-01
        fechaFinal:
          type: string
          description: Fecha final de la inhabilitación. Fecha en formato ISO 8601
          format: date
          example: 2020-12-31
      description: Se llena en caso de que el particular haya sido inhabilitado
      example:
        fechaInicial: 2019-01-01T00:00:00.000+0000
        plazo: 2 años
        fechaFinal: 2020-12-31T00:00:00.000+0000
    resParticularesSancionados_documentos:
      required:
      - descripcion
      - fecha
      - id
      - titulo
      - url
      type: object
      properties:
        id:
          type: string
          description: Puede ser el folio del documento o al menos un consecutivo
            para identificar a los documentos en el registro.
          example: doc-1
        tipo:
          type: string
          description: Tipo de documento, Valores permitidos ["RESOLUCION", "CONSTANCIA_SANCION",
            "CONSTANCIA_INHABILITACION", "CONSTANCIA_ABSTENCION"]
          example: CONSTANCIA_INHABILITACION
          enum:
          - RESOLUCION
          - CONSTANCIA_SANCION
          - CONSTANCIA_INHABILITACION
          - CONSTANCIA_ABSTENCION
        titulo:
          type: string
          description: Titulo del documento
          example: Constancia de inhabilitación...
        descripcion:
          type: string
          description: Descripción del documento
          example: Descripción del documento
        url:
          type: string
          description: Hipervínculo al documento
          example: https://ejemplo.mx/archivo/documento.pdf
        fecha:
          type: string
          description: Fecha del documento. Fecha en formato ISO 8601
          format: date
          example: 2019-11-20
      example:
        descripcion: Descripción del documento
        fecha: 2019-11-20T00:00:00.000+0000
        tipo: CONSTANCIA_INHABILITACION
        titulo: Constancia de inhabilitación...
        id: doc-1
        url: https://ejemplo.mx/archivo/documento.pdf
    resParticularesSancionados_results:
      required:
      - autoridadSancionadora
      - causaMotivoHechos
      - fechaCaptura
      - id
      - institucionDependencia
      - particularSancionado
      - tipoSancion
      type: object
      properties:
        id:
          type: string
          description: Dato que permita identificar de manera  unívoca al registro
          example: "1"
        fechaCaptura:
          type: string
          description: Fecha y hora - en la que se agrega o modifica el registro.
            Fecha en formato ISO 8601
          format: date-time
          example: 2017-07-21T17:32:28Z
        expediente:
          type: string
          description: El expediente derivado de la investigación que las Autoridades
            Investigadoras realizan
          example: ABC123
        institucionDependencia:
          $ref: '#/components/schemas/resParticularesSancionados_institucionDependencia'
        particularSancionado:
          $ref: '#/components/schemas/resParticularesSancionados_particularSancionado'
        objetoContrato:
          type: string
          description: Objeto del contrato
          example: Servicio de telefonía fija
        autoridadSancionadora:
          type: string
          description: OIC o Unidad responsable de la sanción
          example: OIC
        tipoFalta:
          type: string
          description: Tipo de falta. eg.,soborno, participación ilícita, tráfico
            de influencias, uso información falsa,colusión, obstrucción de facultades,
            contratación indebida, uso indebido de recusos públicos.
          example: Contratación indebida
        tipoSancion:
          $ref: '#/components/schemas/tipoSancion'
        causaMotivoHechos:
          type: string
          description: Causa/motivo de la sanción
          example: Causa o motivo de la sanción
        acto:
          type: string
          description: Acto que originó la investigación. Ejm. Contrataciones
          example: Contrataciones
        responsableSancion:
          $ref: '#/components/schemas/resParticularesSancionados_responsableSancion'
        resolucion:
          $ref: '#/components/schemas/resParticularesSancionados_resolucion'
        multa:
          $ref: '#/components/schemas/resParticularesSancionados_multa'
        inhabilitacion:
          $ref: '#/components/schemas/resParticularesSancionados_inhabilitacion'
        observaciones:
          type: string
          description: Cualquier observación pertinente
          example: Se multo e inhabilitó al proveedor PATITO TELECOM S.A de C.V. por
            ...
        documentos:
          type: array
          description: Es una sección que permite adjuntar referencias a cualquier
            documento que se considere de valor para transparentar el proceso de sanción.
          items:
            $ref: '#/components/schemas/resParticularesSancionados_documentos'
      example:
        autoridadSancionadora: OIC
        responsableSancion:
          primerApellido: Pérez
          segundoApellido: López
          nombres: Carlos
        institucionDependencia:
          clave: XYZ987
          siglas: SHCP
          nombre: Secretaría de Hacienda y Crédito Público
        tipoFalta: Contratación indebida
        resolucion:
          sentido: Sansionatoria con multa
          fechaNotificacion: 2019-11-20T00:00:00.000+0000
          url: https://ejemplo.mx/archivo/documento.pdf
        expediente: ABC123
        multa:
          monto: 5000000
          moneda:
            clave: MXN
            valor: Peso Mexicano
        tipoSancion:
        - clave: I
          valor: Inhabilitado
        - clave: M
          valor: Multado
        acto: Contrataciones
        documentos:
        - descripcion: Descripción del documento
          fecha: 2019-11-20T00:00:00.000+0000
          tipo: CONSTANCIA_INHABILITACION
          titulo: Constancia de inhabilitación...
          id: doc-1
          url: https://ejemplo.mx/archivo/documento.pdf
        - descripcion: Descripción del documento
          fecha: 2019-11-20T00:00:00.000+0000
          tipo: CONSTANCIA_INHABILITACION
          titulo: Constancia de inhabilitación...
          id: doc-1
          url: https://ejemplo.mx/archivo/documento.pdf
        fechaCaptura: 2017-07-21T17:32:28Z
        particularSancionado:
          nombreRazonSocial: PATITO TELECOM S.A. de C.V.
          objetoSocial: Objeto social de la empresa
          domicilioExtranjero:
            ciudadLocalidad: Springfield
            codigoPostal: "01234"
            estadoProvincia: North Takoma
            calle: Evergreen Terrace
            numeroExterior: "742"
            numeroInterior: ""
            pais:
              clave: MX
              valor: México
          tipoPersona: F
          directorGeneral:
            primerApellido: Pérez
            segundoApellido: López
            curp: BADD110313HCMLNS09
            nombres: Carlos
          domicilioMexico:
            codigoPostal: "55018"
            municipio:
              clave: "033"
              valor: Ecatepec de Morelos
            entidadFederativa:
              clave: "15"
              valor: México
            localidad:
              clave: "0001"
              valor: Ecatepec de Morelos
            numeroExterior: "24"
            numeroInterior: "48"
            vialidad:
              clave: CALLE
              valor: El Rosal
            pais:
              clave: MX
              valor: México
          apoderadoLegal:
            primerApellido: Pérez
            segundoApellido: López
            curp: BADD110313HCMLNS09
            nombres: Carlos
          telefono: "552123244"
          rfc: PAT21343213
        causaMotivoHechos: Causa o motivo de la sanción
        observaciones: Se multo e inhabilitó al proveedor PATITO TELECOM S.A de C.V.
          por ...
        id: "1"
        objetoContrato: Servicio de telefonía fija
        inhabilitacion:
          fechaInicial: 2019-01-01T00:00:00.000+0000
          plazo: 2 años
          fechaFinal: 2020-12-31T00:00:00.000+0000
    reqParticularesSancionados_sort:
      type: object
      properties:
        nombreRazonSocial:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        rfc:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        institucionDependencia:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
    reqParticularesSancionados_query:
      properties:
        id:
          type: string
          description: Devuelve resultados al buscar por su ID. Si se especifica este
            valor, el objeto query es omitido.
          example: 123abc
        nombreRazonSocial:
          type: string
          description: Nombre o razón social
          example: PATITO TELECOM
        rfc:
          type: string
          description: Se deberá escribir los 10 caracteres básicos y los 3 caracteres
            de la homoclave
          example: PAT21343213
        institucionDependencia:
          type: string
          description: Nombre de la institución o dependencia dónde se cometió la
            falta
          example: Secretaría de Hacienda y Crédito Público
        expediente:
          type: string
          description: El expediente derivado de la investigación que las Autoridades
            Investigadoras realizan
          example: ABC123
        tipoSancion:
          uniqueItems: true
          type: array
          description: "Clave del Tipo de sanción. Valores permitidos\n  * `I` - INHABILITADO\
            \ \n  * `M` - MULTADO\n  * `S` - SUSPENSIÓN DE ACTIVIDADES\n  * `D` -\
            \ DISOLUCIÓN DE LA SOCIEDAD\n  * `A` - AMONESTACIÓN\n  * `IND` - INDEMNIZACIÓ\
            N POR LOS DAÑOS Y PERJUICIOS OCASIONADOS A LA HACIENDA PÚBLICA FEDERAL,\
            \ LOCAL O MUNICIPAL, O AL PATRIMONIO DE LOS ENTES PÚBLICOS.\n  * `SE`\
            \ - SANCIÓN ECONÓMICA\n  * `O` - OTRO\n\n*******Deberá devolver todos\
            \ los registros que contengan al menos uno de los valores solicitados.\n"
          example:
          - I
          - M
          items:
            type: string
            enum:
            - I
            - M
            - S
            - D
            - A
        tipoPersona:
          type: string
          description: 'Los valores pueden ser:  F = Física, M = Moral'
          example: F
      description: En este objeto se pasan los parametros para las busquedas por medio
        de los campos correspondientes
...
  securitySchemes:
    psancionados_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: /oauth
          scopes: {}